Replace global view with machines view.
A "view" of 23,000 nodes isn't a very meaningful concept: in practice, we get an empty white box and a lot of CPU churn. It might be nice to show a rat's nest of "look what you don't have to work with", but it doesn't seem very practical.

What is practical, however, is to show the machines involved in our demo and how they interact with each other.

This closes #1.