Refactor internal/external machine styling.
Move logic to `add_node` (and move that function to graphing.js) so that styling can be applied consistently in both the machine view and the worksheet view. Also use CSS inheritance to simplify styles slightly: mark external nodes as such and treat them specially but let the default style stand for internal nodes.