Tweak machine styling.
Add text outline to machine name (since it overlaps the machine icon) and make external machines just look like computers (no ellipses, boxes or backgrounds, just the machine icon). Also add a machine colour, based on the darkest blue in the icon, for use in the machine font (and, perhaps soon, for sockets that communicate with external machines).