Add further visual tweaks to machines.
Use SVGs for machines, since they tend to get scaled more than other node types. Also, it turns out that Cytoscape has a CSS selector for compound parents, so use this to visually differentiate internal-machines-as-compound-nodes. Finally, use `width` and `height` for non-parent nodes (`min-width` and `min-height` only applies to parent nodes, which are not used in the machines view).