Show files as well as versions in the graph.
When we import a file-version, put it into an abstract "file". This file
node can display the filename(s) once, allowing simplification of the
file-version nodes, and it also makes their relationship more obvious.

It is worth noting that the set of names attached to a file node will
depend on the file-versions that have been added to the graph. For
example, if we add a rename operation to the graph, the abstract file
will acquire another potential name.

Closes issue #15.