Add ability to save worksheet JSON.
By clicking the "Save" button that now appears on the main worksheet, we
can export JSON from Cytoscape that can later be fed back in. This gets
us halfway to resolving #27: I've tested that the JSON export works
with the browser UI security policy and that JSON loading can be done
manually from the JavaScript console. The next step is to integrate JSON
loading in the browser UI parts.