Fix "disappearing buttons" problem.
Put worksheet "Load" and "Save" buttons in a container around the
worksheet rather than in the worksheet itself. This means that Cytoscape
won't erase the buttons when we load a graph (and clear the old one),
and it also has the happy effect of not passing button clicks through to
the graph library.