Work around an issue with too-early Cytoscape creation.
We have been having issues with Cytoscape behaving "weirdly" (e.g.,
nodes disappear when we re-layout) and with Chromium throwing errors on
Cytoscape initialization. This may be because:

* Marko is "mounting" 0-sized components or
* Cytoscape is insufficiently tolerant of 0-sized components or
* Cytoscape has code paths that don't return values (see
github.com/cytoscape/cytoscape.js/issues/2122)

or (most likely) some combination of the three. Whatever the reason, we
ened to work around it by doing a later, post-sizing initialization of
Cytoscape graphs. This unfortunately means JQuery for the moment, though
we may be able to find a less unfortunate workaround in the future.