Increase Neo4j query sophistication.
Rather than passing around an object full of query functions, pass around a function that can return a reference to a database object, and let that object have methods to a) build queries and b) execute them. This makes for a cleaner division of responsibilities, less code duplication and no weird top-level `queries` object.

Along the way, fix a bug in which we were passing the wrong value into `layout.dbConnected()` — the value `this` no longer referred to the originating context, so we need to create a local variable `self=this` that can be captured by the closure.